+++ src/constants.ts
@@ -1,5 +1,4 @@
-export const API_URL = "http://localhost:5007";
-// export const API_URL = "https://api.animaapp.com";
+export const API_URL = "https://api.animaapp.com";
 export const PARAM_KEY = "export-stories";
 export const EVENT_CODE_RECEIVED = "event-code-received";
 export const EXPORT_START = "EXPORT_START";
