--- src/globalStyles.ts
@@ -1,2 +1,0 @@
-export const GLOBAL_STYLES = `
-`;
