+++ src/utils/helpers.ts
@@ -1,0 +1,49 @@
+import { createAlert } from "../components/alert";
+
+export const isString = (value: any) =>
+  Object.prototype.toString.call(value) === "[object String]";
+
+export const isBoolean = (val) => "boolean" === typeof val;
+
+export const capitalize = (s: string) => {
+  if (!isString(s)) return "";
+  return s.charAt(0).toUpperCase() + s.slice(1);
+};
+
+export const createElementFromHTML = (htmlString: string): HTMLElement => {
+  const div = document.createElement("div");
+  div.innerHTML = htmlString.trim();
+  return div.firstElementChild as HTMLElement;
+};
+
+export const notify = (text: string): void => {
+  const alertElement = createAlert(text);
+  document.body.appendChild(alertElement);
+
+  (alertElement.firstElementChild as HTMLElement).style.opacity = "1";
+
+  setTimeout(() => {
+    (alertElement.firstElementChild as HTMLElement).style.opacity = "0";
+    requestAnimationFrame(() => {
+      alertElement.remove();
+    });
+  }, 2500);
+};
+
+export const downloadAsJSON = (data: Record<string, any>) => {
+  const content = JSON.stringify(data, null, 2);
+  const blob = new Blob([content], { type: "text/json" });
+  const dataURI = `data:text/json;charset=utf-8,${content}`;
+  const URL = window.URL || window.webkitURL;
+  const downloadURI =
+    typeof URL.createObjectURL === "undefined"
+      ? dataURI
+      : URL.createObjectURL(blob);
+
+  let link = document.createElement("a");
+  link.setAttribute("href", downloadURI);
+  link.setAttribute("download", "fg-json.json");
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+};
